[ refactor ] : 커뮤니티 패키지 구조 개선 및 도메인별 분리
