--- src/main/java/com/talkit/app/domain/bookMark/entity/BookMark.java
@@ -1,5 +1,0 @@
-package com.talkit.app.domain.bookMark.entity;
-
-public class BookMark {
-
-}
