--- src/main/java/com/talkit/app/domain/bookMark/repository/BookMarkRepository.java
@@ -1,5 +1,0 @@
-package com.talkit.app.domain.bookMark.repository;
-
-public class BookMarkRepository {
-
-}
