+++ src/main/java/com/talkit/app/domain/community/base/dto/CommunityRequestDto.java
@@ -1,0 +1,14 @@
+package com.talkit.app.domain.community.base.dto;
+
+import jakarta.validation.constraints.Size;
+
+import java.util.List;
+
+public record CommunityRequestDto(
+    @Size(max = 100, message = "게시물 제목은 최대 100자까지 입력할 수 있습니다.")
+    String title,
+    String content,
+    String category,
+    List<String> tags
+) {
+}
