+++ src/main/java/com/talkit/app/domain/community/bookMark/repository/BookMarkRepository.java
@@ -1,0 +1,5 @@
+package com.talkit.app.domain.community.bookMark.repository;
+
+public class BookMarkRepository {
+
+}
