+++ src/main/java/com/talkit/app/domain/community/comment/dto/CommunityCommentRequestDto.java
@@ -1,0 +1,11 @@
+package com.talkit.app.domain.community.comment.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record CommunityCommentRequestDto(
+
+    @NotBlank(message = "댓글 내용을 입력해주세요.")
+    String content
+
+) {
+}
