+++ src/main/java/com/talkit/app/domain/community/comment/repository/CommentLikeRepository.java
@@ -1,0 +1,10 @@
+package com.talkit.app.domain.community.comment.repository;
+
+import com.talkit.app.domain.community.entity.CommentLike;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
+
+}
