+++ src/main/java/com/talkit/app/domain/community/comment/service/CommentLikeService.java
@@ -1,0 +1,12 @@
+package com.talkit.app.domain.community.comment.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+@Service
+public class CommentLikeService {
+
+}
