--- src/main/java/com/talkit/app/domain/community/dto/CommunityLikeResponseDto.java
@@ -1,16 +1,0 @@
-package com.talkit.app.domain.community.dto;
-
-import lombok.Builder;
-import lombok.Getter; // 추가
-
-@Getter
-@Builder
-public class CommunityLikeResponseDto {
-    private final Long communityId;
-
-    public static CommunityLikeResponseDto of(Long communityId) {
-        return CommunityLikeResponseDto.builder()
-            .communityId(communityId)
-            .build();
-    }
-}
