+++ src/main/java/com/talkit/app/domain/community/entity/BookMark.java
@@ -1,0 +1,5 @@
+package com.talkit.app.domain.community.entity;
+
+public class BookMark {
+
+}
