+++ src/main/java/com/talkit/app/domain/community/like/dto/CommunityLikeResponseDto.java
@@ -1,0 +1,16 @@
+package com.talkit.app.domain.community.like.dto;
+
+import lombok.Builder;
+import lombok.Getter; // 추가
+
+@Getter
+@Builder
+public class CommunityLikeResponseDto {
+    private final Long communityId;
+
+    public static CommunityLikeResponseDto of(Long communityId) {
+        return CommunityLikeResponseDto.builder()
+            .communityId(communityId)
+            .build();
+    }
+}
