--- src/main/java/com/talkit/app/domain/community/repository/CommentLikeRepository.java
@@ -1,10 +1,0 @@
-package com.talkit.app.domain.community.repository;
-
-import com.talkit.app.domain.community.entity.CommentLike;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
-
-}
