--- src/main/java/com/talkit/app/domain/community/repository/CommunityRepositoryCustom.java
@@ -1,12 +1,0 @@
-package com.talkit.app.domain.community.repository;
-
-import com.talkit.app.domain.community.dto.CommunitySearchConditionDto;
-import com.talkit.app.domain.community.entity.Community;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-
-public interface CommunityRepositoryCustom {
-
-    Page<Community> searchByCondition(CommunitySearchConditionDto condition, Pageable pageable);
-}
