--- src/main/java/com/talkit/app/domain/community/service/CommentLikeService.java
@@ -1,12 +1,0 @@
-package com.talkit.app.domain.community.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-@Service
-public class CommentLikeService {
-
-}
